refactor: move deployment into internal
