+++ packages/core/src/dsl/DeploymentGraph.ts
@@ -2,7 +2,7 @@ import { Graph } from "../graph/Graph";
 import {
   DeploymentGraphVertex as DeploymentGraphVertex,
   ScopeData,
-} from "../types/deploymentGraph";
+} from "../internal/types/deploymentGraph";
 
 export class DeploymentGraph extends Graph<DeploymentGraphVertex> {
   public scopeData: {
