+++ packages/core/src/dsl/buildModule.ts
@@ -1,4 +1,4 @@
-import type { IDeploymentBuilder } from "../types/deploymentGraph";
+import type { IDeploymentBuilder } from "../internal/types/deploymentGraph";
 import type { Module, ModuleDict } from "../types/module";
 
 import { assertStringParam, assertFunctionParam } from "../utils/errors";
