+++ packages/core/src/services/ConfigService.ts
@@ -1,4 +1,4 @@
-import type { ExternalParamValue } from "../types/deploymentGraph";
+import type { ExternalParamValue } from "../internal/types/deploymentGraph";
 import type { HasParamResult, Providers } from "../types/providers";
 
 import { IConfigService } from "../types/services";
