+++ packages/core/src/types/services.ts
@@ -1,4 +1,4 @@
-import type { ExternalParamValue } from "./deploymentGraph";
+import type { ExternalParamValue } from "../internal/types/deploymentGraph";
 import type { Artifact } from "./hardhat";
 import type {
   GasProvider,
