+++ packages/core/src/validation/dispatch/validateEvent.ts
@@ -1,6 +1,6 @@
 import { ethers } from "ethers";
 
-import { EventVertex } from "../../types/deploymentGraph";
+import { EventVertex } from "../../internal/types/deploymentGraph";
 import { VertexResultEnum } from "../../types/graph";
 import {
   ValidationDispatchContext,
