+++ packages/core/src/validation/validateDeploymentGraph.ts
@@ -1,6 +1,9 @@
 import { getSortedVertexIdsFrom } from "../graph/utils";
 import { visit } from "../graph/visit";
-import { CallPoints, IDeploymentGraph } from "../types/deploymentGraph";
+import {
+  CallPoints,
+  IDeploymentGraph,
+} from "../internal/types/deploymentGraph";
 import { Services } from "../types/services";
 import { ValidationVisitResult } from "../types/validation";
 import { IgnitionError } from "../utils/errors";
