+++ packages/core/test/deploymentBuilder/sendETH.ts
@@ -7,7 +7,7 @@ import { generateDeploymentGraphFrom } from "../../src/internal/process/generate
 import {
   IDeploymentBuilder,
   IDeploymentGraph,
-} from "../../src/types/deploymentGraph";
+} from "../../src/internal/types/deploymentGraph";
 import { isHardhatContract } from "../../src/utils/guards";
 
 import {
